### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-16131

### Root cause
- The API return rate as double value in `PriceTierDto` object, we expect `rate` as `int` value so it causes an error when we map data from API to `PriceTierDto`

### Solution
- Change `rate` and `minAmount` field in PriceTierDto to double to avoid mapping error
- Round up the value of the displayed price in `StringUtils`  across the app for the ID market only

### Test account
- environment: UAT
- username: rootadmin
- salesOrg: ID market
- customer: default
- shipto: default
- material/setup: AIABG6

### Output screenshot
![Simulator Screenshot - iPhone 15 - 2023-12-28 at 11.57.34.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/90949/attachments/Simulator%20Screenshot%20-%20iPhone%2015%20-%202023-12-28%20at%2011.57.34.png =x400)

+++ lib/domain/order/entities/price_tier.dart
@@ -37,8 +37,8 @@ class PriceTierItem with _$PriceTierItem {
     required int sequence,
     required int quantity,
     required double rate,
-    required int percentage,
-    required int minAmount,
+    required double percentage,
+    required double minAmount,
     required ScaleBasis scaleBasis,
   }) = _PriceTierItem;
